refactor(cli): rename main-menu.ts to main.ts
