+++ src/cli/commands/main.ts
